Feat: 즐겨찾기한 사진 Swiper 적용
